1. db seeder added 2. post commenting implemented

+++ XamarinBlogEducation.ViewModels/Models/Blog/AddCommentBlogViewModel.cs
@@ -7,7 +7,7 @@ namespace XamarinBlogEducation.ViewModels.Models.Blog
     public class AddCommentBlogViewModel
     {
         public string Content { get; set; }
-        public int PostId { get; set; }
+        public long PostId { get; set; }
         public string UserId { get; set; }
 
     }
